#2619 Fix galaxy selector - broken and slow down deletes with gaps
During DB change, .ID() in galaxy selector was calling the wrong function (the int/128)

To prevent this, change the name of the function to  ID128 and make sure the galaxy selector is using the double one.

This was preventing galaxy selection.

Also slow it down with gaps so it does not hog the db

+++ EDDiscovery/Forms/GalaxySectorSelect.cs
@@ -377,7 +377,7 @@ namespace EDDiscovery.Forms
                 xly = (int)(lypostop.X + fractx * lywidth);
                 zly = (int)(lypostop.Y - fracty * lyheight);
 
-                int id = GridId.Id(xly, zly);
+                int id = GridId.Id((double)xly, (double)zly);
 
                // System.Diagnostics.Debug.WriteLine("ID {0} {1} {2}" ,id, xly, zly);
 
@@ -403,6 +403,8 @@ namespace EDDiscovery.Forms
                         includedgridid.Add(id);
                         includedgridid.Sort();
                     }
+
+                    //System.Diagnostics.Debug.WriteLine("Grid list now " + string.Join(",",includedgridid));
                     Invalidate();
                     onChange?.Invoke();
                 }
